producer/requirements.txt
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pyyaml==6.0.1
python-docx==1.1.2
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
pinecone==6.0.2
openai==1.50.0
tiktoken==0.7.0
